cv-image-processor/requirements.txt
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.43.1
graphviz==0.20.1
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.5.0
jupyter_core==5.5.0
kafka-python==2.0.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
minio==7.2.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.59.1
numpy==1.26.1
onnxruntime==1.17.3
opencv-python-headless==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.1
prompt-toolkit==3.0.39
protobuf==5.26.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.1
pycodestyle==2.10.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.16.1
PyMatting==1.1.12
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.1.1
redis==5.0.4
referencing==0.35.0
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
Telethon==1.34.0
tifffile==2024.4.24
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.9
zipp==3.18.1
